third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEConvolveMatrixElement-dom-in-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' attribute of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.getAttribute('in') is "SourceAlpha"
PASS convolveMatrixElement.getAttribute('in') is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
